refactor: Improve view.html to include external url and action buttons
This commit refactors the view.html template to include the external url of the article as well as action buttons to change the entry status and open the external link. Additionally, the commit improves the entry-meta by displaying the website where the article was published and the entry date.

Also: Add fallback URL for unpublished articles

Return the external domain URL when wanting to publish an article with no publisher. Also, in case the publisher list is none or empty, return the external domain URL.

Also: Refactor article status and bookmarking UI elements

This commit updates the HTML for the UI elements used in changing
an article's status and bookmarking an article. This update makes
the SVG icon and label appear correctly for both actions. The
code also uses a more readable syntax by breaking up the code into
multiple lines.